fix stuck when error
